feat: fix price
